requirements.txt
streamlit
streamlit_metrics
streamlit-webrtc
st-clickable-images
pandas
numpy
seaborn
matplotlib
scikit-learn
opencensus-ext-azure
azure-keyvault
azure-identity
azure-ai-textanalytics
azure-cognitiveservices-speech
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-face
opencv-python
librosa
SoundFile
